#71 propagation inside folder
